Revert "refactor unsqueeze to use axes of Span<i32>"
This reverts commit b3348955c285b86bc2128e71951b9e5cf73e3cd7.

+++ src/operators/tensor/implementations/tensor_fp16x16wide.cairo
@@ -324,7 +324,7 @@ impl FP16x16WTensor of TensorTrait<FP16x16W> {
         core_tensor::squeeze(self, axes)
     }
 
-    fn unsqueeze(self: @Tensor<FP16x16W>, axes: Span<i32>) -> Tensor<FP16x16W> {
+    fn unsqueeze(self: @Tensor<FP16x16W>, axes: Span<usize>) -> Tensor<FP16x16W> {
         core_tensor::unsqueeze(self, axes)
     }
